removes unnecessary EC logic
